fix(tracing): Baggage parsing fails when input is not of type string (#5276)
This patch attempts to fix a bug reported in #5250 where the input parameter for `parseBaggageString` was not a string and hence the `string.split` function call would throw an error. We now check for the type of the input and make the function

* return an empty baggage object if the input is neither a string, nor an array. In this case we log a warning to get more information on what is actually passed in.
* parse the baggage string if the input is a string
* join and then split the array if the input is an array

Additionally, the parser now ignores empty or blank baggage entries (e.g. `foo=bar,   ,,alice=bob`).

Added tests for parsing the new data types