fix(node): Allow `ParseRequestOptions` to be passed to request handler (#5287)
This fixes an issue introduced by https://github.com/getsentry/sentry-javascript/pull/5257, which was designed to be  backwards-compatible, but which missed a key use case: When users pass options to our Express middleware, they might be (and in fact, for now, almost certainly are) passing them in the form of `ParseRequestOptions`, not `AddRequestDataToEventOptions`. This allows for that possibility, as a backwards-compatibility measure until v8. It also adjusts a spot in the serverless SDK where a similar problem can occur.

Fixes https://github.com/getsentry/sentry-javascript/issues/5282.