ref(tracing): Unify DSC key names in envelope and baggage headers (#5302)
Unify the key names in the envelope `trace` header and  the `baggage` Http header. As a result, we get rid of the mapping that was necessary before, thereby loosing a little JS bundle weight ;) 
Furthermore, the PR renames the `EventTraceContext` interface and aligns the `Baggage` type so that we can ensure that both DSC transport mechanisms are based on the same keys.