fix(build): Use Node 14 in "Size Check" job to fix CI runs on master (#5324)
Set "Size Limit" job node version to 14 to fix Ember build failing. Because of how the size-limit-action handles runs on master, it actually builds the SDKs again instead of using the build cache like in PR contexts. With this patch we fix the CI runs on our end to unblock releases and make CI pass on master again