fix(remix): Move hook checks inside the wrapper component. (#5371)
`withSentryRouteTracing` is called on `root.tsx` which runs before `entry.client.tsx` (where `Sentry.init` for browser is called and `remixRouterInstrumentation` is assigned). 

The check if the provided hooks are available was failing silently, and causing `withSentryRouteTracing` to early return the `App` unwrapped.

This PR ensures that validation runs on the client side, after `entry.client.tsx`.