ref(angular): Add transaction source for Angular Router (#5382)
Add the transaction name source annotation to the Angular SDK. Since the Angular SDK currently does not parameterize URLs, we only assign `'url'` as the transaction name source. We're revisiting parameterization in Angular in a follow up PR.

Additionally, add a two tests to cover this change (might be more relevant once we add parameterization).