ref(nextjs): Remove compensation for workaround in `_error.js` (#5378)
For a long time, the `_error.js` page we provide `@sentry/nextjs` users contained a workaround for https://github.com/vercel/next.js/issues/8592. When most of the code in `_error.js` was moved into `captureUnderscoreErrorException`, parts of the workaround came along for the ride. Now that that issue has been fixed, the workaround is no longer necessary.

This removes as many of the workaround's vestiges in `captureUnderscoreErrorException` as possible. (We can't remove them all, because the workaround existed in user code rather than ours, and we can't stop people from continuing to use it. This fixes things so that if they do that, it'll just bail gracefully.)