feat: Capture the correct timestamps and initial URL when sending events to Sentry  (getsentry/sentry-replay#115)
Ensure that we capture the correct initial state (e.g. timestamps + url) when sending the root replay event, as well as replay updates (timestamp only). For the root event, we defer creating + sending the event until after a certain amount of time + events elapses, but the session start point should be when we first start collecting replay events (recordings or otherwise).

For replay updates, the timestamp should be the time when we start flushing the event buffer, but the event should only be created after successfully uploading the replay.

Note backend will need to be able to support the arbitrary starting timestamp as normal event ingestion does not.