feat(tracing): Log start and end of span (#5446)
Right now, if one of our SDKs is running on a cloud provider and therefore all the user has is logs, there's no real way to debug missing spans. Are they not finishing in time and getting dropped? Are they even starting in the first place? No way to know.

This fixes that by adding logging to the start and end of spans., the same way we have with transactions.