ref(tracing): Simplify sample_rate serialization for DSC (#5475)
Changes our formatting logic of the `sample_rate` value when the dynamic sampling context (DSC) is populated. Instead of using `Number.toLocaleString` with a bunch of options (which theoretically is supported by all browsers; but it turns out that IE11 doesn't support the `fullwide` locale we use), we now simply convert the number `toString` to do the job. Given that Relay now supports parsing the sample_rate from JSON format, we can drop our previous formatting logic. 

Deleted a test that's IMO no longer necessary, covering the conversion of exponential to decimal notation.