fix(nextjs): Add default `distDir` value back into `index.server.ts` (#5479)
In https://github.com/getsentry/sentry-javascript/pull/5445, a change was made to the way the global `RewriteFrames` helper value is handled. Specifically, setting the default value (using the `||` operator) was moved to the place where the value is set rather than where it's retrieved. But if something goes wrong and for whatever reason the value never gets set globally, it now causes errors when the value is later used, because it has nothing to default to.

This fixes that by restoring the default value to the old location, so that when it's used, it will never be undefined.

Fixes https://github.com/getsentry/sentry-javascript/issues/5478.