fix: Duplicate events due to compression WebWorker (getsentry/sentry-replay#127)
Soo multiple calls to `flushUpdate()` could result in duplicated events. This is due to the async nature of WebWorkers and our message passing. Calls to `flushUpdate()` were attaching multiple event listeners to the WebWorker before the first WebWorker is able to respond. The listeners had no unique identifiers for the responses, so *all* of the listeners would resolve as soon as it receives a "finished" response from the WebWorkers.

To fix this, add a unique ID to each request/response so that the callers were not listening to the same response from the WebWorker.

Also adds `jsdom-worker` for tests instead of the poorly mocked out tests before.

Closes getsentry/sentry-replay#99