fix(remix): Return response if detected in root loader (#5558)
This patch updates the remix root loader wrapper to return a response when possible. This allows user set headers to be propogated.

We do this by vendoring in Remix's `json` utility and using it accordingly. We also now return the response directly instead of the data in cases where the data is not an object.