ref(test): Refactor Node integration tests to prevent nock leaks. (#5579)
Currently, Node integration tests initialize nock interceptors on server initialization and end them after a certain timeout. This recently created problems with asynchronous event processors, and caused random cases of envelope leaks between tests.

To resolve it, this approach returns `http.Server` and `nock.Scope` instances back to the test, to eventually be provided to a helper that either collects envelopes or makes requests. The responsibility of ending those two instances is moved to the collection helpers like `getEnvelopeRequest` or `getAPIResponse`.

Tests that do not use those helpers should close the server and nock interceptors manually. 
This way, we ensure that we do not lose the context of the server and the nock interceptors between tests.