fix(remix): Use domains to prevent scope bleed (#5570)
This patch wraps both the express and built-in request handlers with domains, preventing scope bleeding from happening. It also updates our express request handler to match that of the `withSentry` handler from NextJS. In a follow-up patch, we'll consolidate the implementations of both these handlers. To test this, we've added an integration test.

Co-authored-by: Onur Temizkan <onur@narval.co.uk>