ref(tests): Refactor Node integration tests further. (#5596)
It seems that whenever a manual `http` request is initiated from a test (or test server) without waiting for / asserting on events or transactions, there are also transactions as side effects which may or may not finish before the test ends. This specifically happens with tests that assert request headers for `trace` and `baggage`, because they don't need to wait for the request / scenario to finish.

So it looks like the leakage to the next test doesn't come from the `nock` interceptor or an unclosed server, it comes from an unflushed Sentry transaction. The other tests using `getEnvelopeRequest` or `getMultipleEnvelopeRequest` finish wait for the events to be flushed.

While debugging this, I also made some more refactors in utilities, which I think will make them more concise.

Summary:

- [x] Making sure the events are flushed before tests that use `getAPIResponse` or `runScenario` complete.
- [x] Converted `http` usages in helpers to `axios` calls for simplicity and debuggability.
- [x] Moved `nock` initialization out of `runServer` so we don't need to carry `nock.Scope` around anymore. The `nock` instance is either manually initialized, or inside the `getMultipleEnvelopeRequest` function. The tests that are using manual `http` calls don't normally need nock anyways.
- [x] Filtering envelopes by their types on `nock` level, so if there is a need to filter them, we don't lose the expected `count`. Also, that makes the envelope indices inside tests simpler.
- [x] Switched to `nock.cleanAll()` and `nock.removeInterceptor(mock)` to end `nock` instead of `.persist(false)`.

Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>